+++ airbyte-cdk/python/airbyte_cdk/destinations/vector_db_based/indexer.py
@@ -25,6 +25,9 @@ class Indexer(ABC):
     def pre_sync(self, catalog: ConfiguredAirbyteCatalog) -> None:
         """
         Run before the sync starts. This method should be used to make sure all records in the destination that belong to streams with a destination mode of overwrite are deleted.
+
+        Each record has a metadata field with the name airbyte_cdk.destinations.vector_db_based.document_processor.METADATA_STREAM_FIELD which can be used to filter documents for deletion.
+        Use the airbyte_cdk.destinations.vector_db_based.utils.create_stream_identifier method to create the stream identifier based on the stream definition to use for filtering.
         """
         pass