+++ airbyte-ci/connectors/pipelines/pipelines/airbyte_ci/connectors/commands.py
@@ -4,18 +4,12 @@
 
 import os
 from pathlib import Path
-from typing import List, Set, Tuple
+from typing import List, Optional, Set, Tuple
 
-import click
+import asyncclick as click
 from connector_ops.utils import ConnectorLanguage, SupportLevelEnum, get_all_connectors_in_repo
 from pipelines import main_logger
-from pipelines.airbyte_ci.connectors.build_image.commands import build
-from pipelines.airbyte_ci.connectors.bump_version.commands import bump_version
-from pipelines.airbyte_ci.connectors.list.commands import list
-from pipelines.airbyte_ci.connectors.migrate_to_base_image.commands import migrate_to_base_image
-from pipelines.airbyte_ci.connectors.publish.commands import publish
-from pipelines.airbyte_ci.connectors.test.commands import test
-from pipelines.airbyte_ci.connectors.upgrade_base_image.commands import upgrade_base_image
+from pipelines.cli.lazy_group import LazyGroup
 from pipelines.helpers.connectors.modifed import ConnectorWithModifiedFiles, get_connector_modified_files, get_modified_connectors
 
 ALL_CONNECTORS = get_all_connectors_in_repo()
@@ -95,7 +89,7 @@ def get_selected_connectors_with_modified_files(
     return selected_connectors_with_modified_files
 
 
-def validate_environment(is_local: bool, use_remote_secrets: bool):
+def validate_environment(is_local: bool):
     """Check if the required environment variables exist."""
     if is_local:
         if not Path(".git").is_dir():
@@ -105,18 +99,64 @@ def validate_environment(is_local: bool, use_remote_secrets: bool):
             "GCP_GSM_CREDENTIALS",
             "CI_REPORT_BUCKET_NAME",
             "CI_GITHUB_ACCESS_TOKEN",
+            "DOCKER_HUB_USERNAME",
+            "DOCKER_HUB_PASSWORD",
         ]
         for required_env_var in required_env_vars_for_ci:
             if os.getenv(required_env_var) is None:
                 raise click.UsageError(f"When running in a CI context a {required_env_var} environment variable must be set.")
-    if use_remote_secrets and os.getenv("GCP_GSM_CREDENTIALS") is None:
-        raise click.UsageError(
-            "You have to set the GCP_GSM_CREDENTIALS if you want to download secrets from GSM. Set the --use-remote-secrets option to false otherwise."
-        )
 
 
-@click.group(help="Commands related to connectors and connector acceptance tests.")
-@click.option("--use-remote-secrets", default=True)  # specific to connectors
+def should_use_remote_secrets(use_remote_secrets: Optional[bool]) -> bool:
+    """Check if the connector secrets should be loaded from Airbyte GSM or from the local secrets directory.
+
+    Args:
+        use_remote_secrets (Optional[bool]): Whether to use remote connector secrets or local connector secrets according to user inputs.
+
+    Raises:
+        click.UsageError: If the --use-remote-secrets flag was provided but no GCP_GSM_CREDENTIALS environment variable was found.
+
+    Returns:
+        bool: Whether to use remote connector secrets (True) or local connector secrets (False).
+    """
+    gcp_gsm_credentials_is_set = bool(os.getenv("GCP_GSM_CREDENTIALS"))
+    if use_remote_secrets is None:
+        if gcp_gsm_credentials_is_set:
+            main_logger.info("GCP_GSM_CREDENTIALS environment variable found, using remote connector secrets.")
+            return True
+        else:
+            main_logger.info("No GCP_GSM_CREDENTIALS environment variable found, using local connector secrets.")
+            return False
+    if use_remote_secrets:
+        if gcp_gsm_credentials_is_set:
+            main_logger.info("GCP_GSM_CREDENTIALS environment variable found, using remote connector secrets.")
+            return True
+        else:
+            raise click.UsageError("The --use-remote-secrets flag was provided but no GCP_GSM_CREDENTIALS environment variable was found.")
+    else:
+        main_logger.info("Using local connector secrets as the --use-local-secrets flag was provided")
+        return False
+
+
+@click.group(
+    cls=LazyGroup,
+    help="Commands related to connectors and connector acceptance tests.",
+    lazy_subcommands={
+        "build": "pipelines.airbyte_ci.connectors.build_image.commands.build",
+        "test": "pipelines.airbyte_ci.connectors.test.commands.test",
+        "list": "pipelines.airbyte_ci.connectors.list.commands.list",
+        "publish": "pipelines.airbyte_ci.connectors.publish.commands.publish",
+        "bump_version": "pipelines.airbyte_ci.connectors.bump_version.commands.bump_version",
+        "migrate_to_base_image": "pipelines.airbyte_ci.connectors.migrate_to_base_image.commands.migrate_to_base_image",
+        "upgrade_base_image": "pipelines.airbyte_ci.connectors.upgrade_base_image.commands.upgrade_base_image",
+    },
+)
+@click.option(
+    "--use-remote-secrets/--use-local-secrets",
+    help="Use Airbyte GSM connector secrets or local connector secrets.",
+    type=bool,
+    default=None,
+)
 @click.option(
     "--name",
     "names",
@@ -171,10 +211,24 @@ def validate_environment(is_local: bool, use_remote_secrets: bool):
     default=True,
     type=bool,
 )
+@click.option(
+    "--docker-hub-username",
+    help="Your username to connect to DockerHub.",
+    type=click.STRING,
+    required=False,
+    envvar="DOCKER_HUB_USERNAME",
+)
+@click.option(
+    "--docker-hub-password",
+    help="Your password to connect to DockerHub.",
+    type=click.STRING,
+    required=False,
+    envvar="DOCKER_HUB_PASSWORD",
+)
 @click.pass_context
-def connectors(
+async def connectors(
     ctx: click.Context,
-    use_remote_secrets: bool,
+    use_remote_secrets: Optional[bool],
     names: Tuple[str],
     languages: Tuple[ConnectorLanguage],
     support_levels: Tuple[str],
@@ -186,16 +240,20 @@ def connectors(
     enable_dependency_scanning: bool,
     use_local_cdk: bool,
     enable_report_auto_open: bool,
+    docker_hub_username: str,
+    docker_hub_password: str,
 ):
     """Group all the connectors-ci command."""
-    validate_environment(ctx.obj["is_local"], use_remote_secrets)
+    validate_environment(ctx.obj["is_local"])
 
     ctx.ensure_object(dict)
-    ctx.obj["use_remote_secrets"] = use_remote_secrets
+    ctx.obj["use_remote_secrets"] = should_use_remote_secrets(use_remote_secrets)
     ctx.obj["concurrency"] = concurrency
     ctx.obj["execute_timeout"] = execute_timeout
     ctx.obj["use_local_cdk"] = use_local_cdk
     ctx.obj["open_report_in_browser"] = enable_report_auto_open
+    ctx.obj["docker_hub_username"] = docker_hub_username
+    ctx.obj["docker_hub_password"] = docker_hub_password
     ctx.obj["selected_connectors_with_modified_files"] = get_selected_connectors_with_modified_files(
         names,
         support_levels,
@@ -207,12 +265,3 @@ def connectors(
         enable_dependency_scanning,
     )
     log_selected_connectors(ctx.obj["selected_connectors_with_modified_files"])
-
-
-connectors.add_command(build)
-connectors.add_command(bump_version)
-connectors.add_command(list)
-connectors.add_command(migrate_to_base_image)
-connectors.add_command(publish)
-connectors.add_command(test)
-connectors.add_command(upgrade_base_image)