+++ airbyte-integrations/connectors/source-bing-ads/unit_tests/test_source.py
@@ -2,21 +2,26 @@
 # Copyright (c) 2023 Airbyte, Inc., all rights reserved.
 #
 
-import json
 from unittest.mock import patch
 
 import pytest
 import source_bing_ads
 from airbyte_cdk.models import SyncMode
-from source_bing_ads.source import AccountPerformanceReportMonthly, Accounts, AdGroups, Ads, Campaigns, SourceBingAds
+from source_bing_ads.source import SourceBingAds
+from source_bing_ads.streams import AccountPerformanceReportMonthly, Accounts, AdGroups, Ads, Campaigns
 
 
 @pytest.fixture(name="config")
 def config_fixture():
     """Generates streams settings from a config file"""
-    CONFIG_FILE = "secrets/config.json"
-    with open(CONFIG_FILE, "r") as f:
-        return json.loads(f.read())
+    return {
+        "tenant_id": "common",
+        "developer_token": "fake_developer_token",
+        "refresh_token": "fake_refresh_token",
+        "client_id": "fake_client_id",
+        "reports_start_date": "2020-01-01",
+        "lookback_window": 0,
+    }
 
 
 @pytest.fixture(name="logger_mock")
@@ -27,7 +32,7 @@ def logger_mock_fixture():
 @patch.object(source_bing_ads.source, "Client")
 def test_streams_config_based(mocked_client, config):
     streams = SourceBingAds().streams(config)
-    assert len(streams) == 29
+    assert len(streams) == 56
 
 
 @patch.object(source_bing_ads.source, "Client")